Update David's Code
My route is nearly finished, option and scenario wise, but it is still in need of bug testing, I will see what I can do tonight. (Update) I finished the options and scenarios in my route, all that is needed is to bug test it as a group and merge the two branches, mine and the master branch into one.